find inlier of sif and triangles
given (feature, descriptor, path-of-2d-vertices-coordinates, path-of-3d-vertices-coordinates, path-of-3d-triangles), reture (sf,sd, tri), with sf a  submatrix of feature, sd a submatrix of descriptor, tri a matrix of 18*N, each column is (3 indexs of vertices, 6 values of coordinates of them in 2d, 9 values of coordinates of them in 3d). They have the same number of columns and they are all the SIFT features inside the box.